+++ src/main/java/be/ugent/rml/target/DirectHttpRequestTarget.java
@@ -9,8 +9,8 @@ import java.util.Map;
 
 public class DirectHttpRequestTarget extends HttpRequestTarget {
 
-    public DirectHttpRequestTarget(Map<String, String> httpRequestInfo, String serializationFormat, List<Quad> metadata, HttpRequestTargetHelper httpRequestTargetHelper) throws IOException {
-        super(httpRequestInfo, serializationFormat, metadata, httpRequestTargetHelper);
+    public DirectHttpRequestTarget(Map<String, String> httpRequestInfo, Map<String, String> httpRequestHeaders, String serializationFormat, List<Quad> metadata, HttpRequestTargetHelper httpRequestTargetHelper) throws IOException {
+        super(httpRequestInfo, httpRequestHeaders, serializationFormat, metadata, httpRequestTargetHelper);
     }
 
     @Override
@@ -32,7 +32,7 @@ public class DirectHttpRequestTarget extends HttpRequestTarget {
     public void close() {
         super.close();
         try {
-            this.httpRequestTargetHelper.executeHttpRequest(httpRequestInfo);
+            this.httpRequestTargetHelper.executeHttpRequest(httpRequestInfo, httpRequestHeaders);
         } catch (Exception e) {
             logger.error("Failed to close http request target for {}: {}",
                     this.httpRequestInfo.get("absoluteURI"), e.getMessage());